Got speed power up and speed reset on max damage done and working
Next is to set up shield alpha on and fade on hit

+++ Assets/---------------Scripts------------/---------------UI---------------/LifeBar.cs
@@ -7,13 +7,13 @@ public class LifeBar : MonoBehaviour
 {
     // the worlds most simple life bar slider script O_O - a bit buggy right now.
     public Slider slider;
-    public void SetMaxLife(int life)
+    public void SetMaxLife(float life)
     {
         slider.maxValue = life;
         slider.value = life;
     }
 
-    public void SetLife(int life)
+    public void SetLife(float life)
     {
         slider.value = life;
     }
